refactor(controller): split selection edits module
